move grinder code from mod init (`grinder/init.lua`) into package init (`grinder/src/grinder.lua`)
